compose: Inline send_request_to_schedule_message.
